+++ [COMP345] Assignment 1.vcxproj
@@ -148,8 +148,8 @@
     <ClInclude Include="ProjectSourceCode\demoCharacter.h" />
     <ClInclude Include="ProjectSourceCode\item.h" />
     <ClInclude Include="ProjectSourceCode\itemcontainer.h" />
-    <ClInclude Include="ProjectSourceCode\Observable.h" />
-    <ClInclude Include="ProjectSourceCode\Observer.h" />
+    <ClInclude Include="ProjectSourceCode\Observer\Observable.h" />
+    <ClInclude Include="ProjectSourceCode\Observer\Observer.h" />
     <ClInclude Include="ProjectSourceCode\testCharacter.h" />
     <ClInclude Include="ProjectSourceCode\testDice.h" />
     <ClInclude Include="ProjectSourceCode\testitem.h" />
